+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## [Unreleased][unreleased]
 - Use money parser for donation amount.
 
+## [2.1.1] - 2020-04-03
+- Fixed incorrect currency symbol filter.
+- Set plugin integration name.
+
 ## [2.1.0] - 2020-03-19
 - Extension extends abstract plugin integration.
 
@@ -66,7 +70,8 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## 1.0.0 - 2015-11-05
 - First release.
 
-[unreleased]: https://github.com/wp-pay-extensions/charitable/compare/2.1.0...HEAD
+[unreleased]: https://github.com/wp-pay-extensions/charitable/compare/2.1.1...HEAD
+[2.1.1]: https://github.com/wp-pay-extensions/charitable/compare/2.1.0...2.1.1
 [2.1.0]: https://github.com/wp-pay-extensions/charitable/compare/2.0.4...2.1.0
 [2.0.4]: https://github.com/wp-pay-extensions/charitable/compare/2.0.3...2.0.4
 [2.0.3]: https://github.com/wp-pay-extensions/charitable/compare/2.0.2...2.0.3